fix(profile): refactor UI profile
