+++ src/components/Profile/FormUserProfile.css
@@ -1,0 +1,101 @@
+.user {
+    display: flex;
+    color: black;
+    line-height:2;
+    font-size: 20px;
+}
+
+.profile-details{
+    width: 100%;
+    margin-left: 20px;
+}
+
+.user-profile {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background: white;
+    padding: 5px;
+    border-radius: 10px;
+    margin-bottom: 5px;
+}
+
+.user-buttons {
+    display: flex;
+    padding-top: 20px;
+    justify-content: space-around;
+}
+
+.user-button {
+    width: 100px;
+    height: 40px;
+    font-size: 1rem;
+    color: black;
+    background-color: #ffd179;
+    font-weight: 600;
+    border: none;
+    cursor: pointer;
+    border-radius: 10px;
+}
+
+.user-button:hover {
+    background-color: #ffb86c;
+    color:white
+}
+
+.user-avatar {
+    width: 120px;
+    border-right: 1px solid #282c2d;
+}
+
+.avatar {
+    width: 100px;
+    height: 100px;
+    padding-left: 5px;
+}
+
+.user-change-avatar {
+    visibility: hidden;
+}
+
+.user-button-change {
+    display: block;
+    width: 100px;
+    font-size: 15px;
+    padding: 5px;
+    background: #ffd179;
+    cursor: pointer;
+    border-radius: 10px;
+}
+
+.user-button-change:hover{
+    background-color: #ffb86c;
+    color:white
+}
+
+.input-name{
+    width: 77%;
+}
+
+.input-email{
+    width: 86%;
+}
+
+.input-phone{
+    width: 70%;
+}
+
+.input-address{
+    width: 83%;
+}
+
+.input-radio {
+    margin-right: 5px;
+}
+
+.input-birthday {
+    background: none;
+    border: none;
+    font-size: 18px;
+    width: 50%;
+}
