--- src/components/Profile/UserProfile.css
@@ -1,65 +1,0 @@
-.user {
-    color: black;
-    line-height:2;
-    font-size: 20px;
-}
-
-.user-profile {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.user-button {
-    padding: 10px 20px;
-    font-size: 1rem;
-    margin-top: 5%;
-    width: 100%;
-    color: black;
-    background-color: #ffd179;
-    font-weight: 600;
-    border: none;
-    cursor: pointer;
-    border-radius: 10px;
-}
-
-.user-button:hover {
-    background-color: #ffb86c;
-    color:white
-}
-
-
-.user-input {
-    border: none;
-    border-bottom: 1px solid black;
-    background: none;
-    outline: none;
-    font-size: 18px
-}
-
-.input-name{
-    width: 77%;
-}
-
-.input-email{
-    width: 86%;
-}
-
-.input-phone{
-    width: 68%;
-}
-
-.input-address{
-    width: 81%;
-}
-
-.input-radio {
-    margin-right: 5px;
-}
-
-.input-birthday {
-    background: none;
-    border: none;
-    font-size: 18px;
-    width: 50%;
-}
